Remove tests against invalid syntax
Rubocop doesn't actually run cops against invalid syntax. These passed trivially
with rubocop 0.52 and earlier, but due to an implementation change in the
methods we call to run tests they fail in 0.53. In the command line, rubocop
would have already bailed before that method is invoked, so this is pointless.